refactor: extract runtime resume safety
