+++ lensnode/lensnode/agent_runtime/runtime.py
@@ -33,7 +33,6 @@ from ..agent_tools import (
     build_general_chat_tools,
 )
 from ..checkpoint import (
-    CheckpointResumeError,
     checkpoint_enabled,
     get_checkpoint_saver,
     load_resume_state,
@@ -66,6 +65,10 @@ from .prompts import (
     history_artifact_guidance as _history_artifact_guidance,
     pick_text as _pick_text,
 )
+from .resume import (
+    pending_checkpoint_tool_calls as _pending_checkpoint_tool_calls,
+    reject_unsafe_resume_tool_replay as _reject_unsafe_resume_tool_replay,
+)
 from ..runtime_resources import (
     cleanup_runtime_resources,
     prepare_runtime_resources,
@@ -87,98 +90,6 @@ class EmptyAgentResponseError(RuntimeError):
     """The agent and its one recovery attempt both returned no text."""
 
     code = "EMPTY_AGENT_RESPONSE"
-
-
-_POTENTIALLY_SIDE_EFFECTING_TOOLS = {
-    "call_skill_api",
-    "run_skill_artifact",
-    "run_skill_script",
-    "run_skill_transform",
-    "save_deliverable",
-}
-
-
-def _pending_checkpoint_tool_calls(
-    messages,
-    pending_write_tool_call_ids=(),
-):
-    """Return tool calls whose result is absent from checkpoint messages."""
-
-    completed = {
-        str(tool_call_id) for tool_call_id in pending_write_tool_call_ids
-    }
-    for message in reversed(tuple(messages or ())):
-        if isinstance(message, ToolMessage):
-            tool_call_id = getattr(message, "tool_call_id", None)
-            if tool_call_id:
-                completed.add(str(tool_call_id))
-            continue
-        tool_calls = getattr(message, "tool_calls", None) or []
-        pending = [
-            call
-            for call in tool_calls
-            if str(call.get("id") or "") not in completed
-        ]
-        if pending:
-            return pending
-    return []
-
-
-def _reject_unsafe_resume_tool_replay(
-    messages,
-    tools,
-    pending_write_tool_call_ids=(),
-):
-    """Fail closed if resume could repeat an external write operation."""
-
-    tools_by_name = {
-        str(getattr(tool, "name", "") or ""): tool for tool in tools or []
-    }
-    for message in tuple(messages or ()):
-        for call in getattr(message, "tool_calls", None) or []:
-            arguments = call.get("args") or {}
-            if (
-                call.get("name") == "call_skill_api"
-                and isinstance(arguments, dict)
-                and arguments.get("capture")
-            ):
-                raise CheckpointResumeError(
-                    "Cannot resume a checkpoint that depends on ephemeral "
-                    "Skill API session values."
-                )
-    for call in _pending_checkpoint_tool_calls(
-        messages,
-        pending_write_tool_call_ids,
-    ):
-        name = str(call.get("name") or "")
-        tool = tools_by_name.get(name)
-        if tool is None:
-            raise CheckpointResumeError(
-                "Cannot resume a checkpoint with an unclassified pending "
-                "tool operation."
-            )
-        metadata = getattr(tool, "metadata", None) or {}
-        if not isinstance(metadata, dict):
-            metadata = {}
-        if metadata.get("idempotent") is True:
-            continue
-        if (
-            metadata.get("read_only") is True
-            or metadata.get("operation") == "read"
-        ):
-            continue
-        is_write = (
-            metadata.get("operation") == "write"
-            or metadata.get("read_only") is False
-            or metadata.get("side_effects") is True
-            or name.startswith("mcp__")
-            or name in _POTENTIALLY_SIDE_EFFECTING_TOOLS
-        )
-        if is_write:
-            raise CheckpointResumeError(
-                "Cannot resume a checkpoint with a pending non-idempotent "
-                "write operation."
-            )
 
 
 class TraceObservationMiddleware(AgentMiddleware):
